% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{a vector of values}

\item{fun}{a string with the name of the value}

\item{alpha}{the value of alpha for the percentage calculation of CI}

\item{cx}{default cx value is 1.5 scale}

\item{...}{edits the histogram}
}
\value{
a list with the given CI and other values
}
\description{
Graphs the bootstrap confidence interval of the input of choice
}
\details{
This is a function created in the Math4753 package that when given parameters of alpha and the given statistic it will graph the confidence interval using the bootstrap method
}
\examples{
myboot2(10000,x=sam,fun="var",alpha=0.20,xlab="var(x)",col="Green",cx=1.5)
#
}
