% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
confidence interval of x
}
\description{
returns a ci for the given sample set
}
\details{
when given a sample vector of values (x) this function will return a confidence interval constructed from a ttest
}
\examples{
myci(x)

}
