% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{number}

\item{iter}{amount}

\item{a}{the starting value of the uniform sample}

\item{x}{the ending value of the uniform sample}
}
\value{
graphical output of the CLT tests
}
\description{
graphical function using the CLT
}
\details{
This function takes in an n,iter,a,b and returns a distribution and a graph using the central limit theorem
}
\examples{
myclt(n=5,iter=10000,a=0,b=5)
}
