% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{myf}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a list of cubed and squared vectors
}
\description{
returns the squares and cubes of the given vector
}
\details{
This is a function created in the Math4753 package that when given a vector it returns a list of squares and cubes
}
\examples{
x=1:4;myf(x)
}
