% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a function}

\item{x}{vetor of param values}
}
\value{
a list with the following values list(i=i,parami=param[i],yi=y[i],slope=slope)
}
\description{
Finds the maximum likelihood estimate
}
\details{
This is a function created in the Math4753 package that when given a function, param and x it gives you the Maximum Likelihood Estimate
}
\examples{
mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),lfun=logbin,main="Binomial",cex.main=1,xlab=expression(p))
#
}
