% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{mynbin}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{size}

\item{r}{rth success}

\item{p}{probability}
}
\value{
The Negative binomial distribution
}
\description{
Calculates negative binomial distribution
}
\details{
when given a sample size, a rth success, and a probability, this function gives the negative binomial distribution
}
\examples{
mybin(10, 3, .4)
}
