% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{x = a (-inf < y < a)}
}
\value{
graphical output
}
\description{
displays a curve and an area
}
\details{
when given a mean, sd, and an a value, this function graphs the normal distribution and area between -infinity and a
}
\examples{
myncurve(10, 5, 6)
}
