% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc param}

\item{xmax}{maximum x}

\item{n}{number of values}

\item{alpha}{the value of alpha}
}
\value{
a list with the q and p-value
}
\description{
When given a t0 it calculates alpha/2 and graphs the distribution
}
\details{
This is a function created in the Math4753 package that when given parameters of alpha and t0 it will graph the p-value and display p-values areas
}
\examples{
mypvalue=function(t0,xmax=4,n=20, alpha=0.05)
#
}
